backend/requirements.txt
Flask
flask-mysql-connector
Flask-CORS